+ 登陆

+++ src/player/component/player_state.tpp
@@ -7,7 +7,7 @@
 template <typename T>
 std::shared_ptr<T> celeritas::player_state::get_component() const
 {
-    const auto component = components_.at(T::get_player_component_type());
+    const auto component = components_.at(static_cast<int>(T::get_player_component_type()));
 
     return boost::polymorphic_pointer_cast<T>(component);
 }
